test: Add structure for recording bad error messages
Notes inline. I can add some links from our contributing docs. I think there's lots to do here (check out the first example for the error message quality...), and it might be a good place for lots of smaller contributions.
