+++ prql-compiler/src/tests/mod.rs
@@ -1,0 +1,4 @@
+mod test;
+mod test_bad_error_messages;
+
+pub use test::compile;
